refactor(connlib): combine pending authorization tracking
